test(iceberg-rest): drop H2-only purge store tests duplicated by backend matrix
The TestIcebergPurgeJobStore + PurgeTestBackend + PurgeTestSchema setup
re-tested IcebergPurgeJobStore against a hand-rolled H2 mock, duplicating
what TestIcebergPurgeJobStoreBackend already covers via the real
TestJDBCBackend matrix (H2 by default, MySQL/PostgreSQL under dockerTest).
Fold the two unique scenarios (markFailed terminal path and the extra
hasActiveJob assertions) into AbstractIcebergPurgeJobStoreBackendTest and
remove the redundant H2-only scaffolding.
